Scripts that reduces the VLA radio interferometers data.
Python scripts that reduces the VLA data in B and C configurations in L
band. The VLA CASA pipeline has already processed the data and I
continue to extra flagging, weighting and Self-calibration. I also do
the polarization calibration and solve for the instrumental leakage
term and the Cross terms.